requirements.txt
streamlit
primer3-py